requirements.txt
Django==1.11.2
MySQL-python==1.2.5
PyMySQL==0.7.11
PyYAML==3.12
argparse==1.2.1
certifi==2017.11.5
chardet==3.0.4
diff-match-patch==20121119
django-ckeditor==5.2.2
django-colorful==1.2
django-flat-responsive==1.4.1
django-import-export==0.5.1
django-redis-cache==1.7.1
django-simple-history==1.8.2
djangoql==0.8.5
enum34==1.1.6
et-xmlfile==1.0.1
google-api-python-client==1.6.5
httplib2==0.10.3
idna==2.6
jdcal==1.3
mysqlclient==1.3.10
oauth2client==4.1.2
odfpy==1.3.5
openpyxl==2.4.8
ply==3.10
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygsheets==1.1.3
pytz==2017.2
redis==2.10.5
requests==2.18.4
rsa==3.4.2
six==1.11.0
tablib==0.11.5
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
wsgiref==0.1.2
xlrd==1.0.0
xlwt==1.2.0
